Automatically Merged using SST Pull Request AutoTester
PR Title: Update for Vanadis to correct potential bugs in immediate offsets for branches and jumps 
PR Author: nmhamster

+++ src/sst/elements/vanadis/inst/vslli.h
@@ -61,7 +61,7 @@ public:
 		case VANADIS_FORMAT_INT32:
 			{
 				const uint32_t src_1 = regFile->getIntReg<uint32_t>( phys_int_regs_in[0] );
-                                regFile->setIntReg<uint64_t>( phys_int_regs_out[0], vanadis_sign_extend( (src_1) << imm_value ) );
+                                regFile->setIntReg<uint32_t>( phys_int_regs_out[0], vanadis_sign_extend( (src_1) << static_cast<uint32_t>(imm_value) ) );
 			}
 			break;
 		case VANADIS_FORMAT_FP32:
